requirements.txt
python-dotenv
py-cord --pre
